+++ packages/eas-cli/src/commands/update/index.ts
@@ -4,7 +4,7 @@ import { Errors, Flags } from '@oclif/core';
 import chalk from 'chalk';
 import nullthrows from 'nullthrows';
 
-import { getExpoWebsiteBaseUrl } from '../../api';
+// import { getExpoWebsiteBaseUrl } from '../../api';
 import { ensureBranchExistsAsync } from '../../branch/queries';
 import { transformFingerprintSource } from '../../build/graphql';
 import { ensureRepoIsCleanAsync } from '../../build/utils/repository';
@@ -24,7 +24,7 @@ import {
   UpdateRolloutInfoGroup,
 } from '../../graphql/generated';
 import { PublishMutation } from '../../graphql/mutations/PublishMutation';
-import { AppQuery } from '../../graphql/queries/AppQuery';
+// import { AppQuery } from '../../graphql/queries/AppQuery';
 import Log, { learnMore, link } from '../../log';
 import { ora } from '../../ora';
 import { RequestedPlatform } from '../../platform';
@@ -38,7 +38,7 @@ import {
   buildUnsortedUpdateInfoGroupAsync,
   collectAssetsAsync,
   filterCollectedAssetsByRequestedPlatforms,
-  findCompatibleBuildsAsync,
+  // findCompatibleBuildsAsync,
   generateEasMetadataAsync,
   getBranchNameForCommandAsync,
   getRuntimeToPlatformsAndFingerprintInfoMappingFromRuntimeVersionInfoObjects,
@@ -594,11 +594,11 @@ export default class UpdatePublish extends EasCommand {
 
     Log.addNewLineIfNone();
 
-    const runtimeToCompatibleBuilds = await Promise.all(
-      runtimeToPlatformsAndFingerprintInfoAndFingerprintSourceMapping.map(obj =>
-        findCompatibleBuildsAsync(graphqlClient, projectId, obj)
-      )
-    );
+    // const runtimeToCompatibleBuilds = await Promise.all(
+    //   runtimeToPlatformsAndFingerprintInfoAndFingerprintSourceMapping.map(obj =>
+    //     findCompatibleBuildsAsync(graphqlClient, projectId, obj)
+    //   )
+    // );
 
     for (const runtime of uniqBy(
       runtimeToPlatformsAndFingerprintInfoMapping,
@@ -672,52 +672,56 @@ export default class UpdatePublish extends EasCommand {
         Log.addNewLineIfNone();
       }
 
-      const fingerprintsWithoutCompatibleBuilds = runtimeToCompatibleBuilds.find(
-        ({ runtimeVersion }) => runtimeVersion === runtime.runtimeVersion
-      )?.fingerprintInfoGroupWithCompatibleBuilds;
-      if (fingerprintsWithoutCompatibleBuilds) {
-        const missingBuilds = Object.entries(fingerprintsWithoutCompatibleBuilds).filter(
-          ([_platform, fingerprintInfo]) => !fingerprintInfo.build
-        );
-        if (missingBuilds.length > 0) {
-          const project = await AppQuery.byIdAsync(graphqlClient, projectId);
-          Log.warn('No compatible builds found for the following fingerprints:');
-          for (const [platform, fingerprintInfo] of missingBuilds) {
-            const fingerprintUrl = new URL(
-              `/accounts/${project.ownerAccount.name}/projects/${project.slug}/fingerprints/${fingerprintInfo.fingerprintHash}`,
-              getExpoWebsiteBaseUrl()
-            );
-            Log.warn(
-              formatFields(
-                [
-                  {
-                    label: `${this.prettyPlatform(platform)} fingerprint`,
-                    value: fingerprintInfo.fingerprintHash,
-                  },
-                  { label: 'URL', value: fingerprintUrl.toString() },
-                ],
-                {
-                  labelFormat: label => `    ${chalk.dim(label)}:`,
-                }
-              )
-            );
-            Log.addNewLineIfNone();
-          }
-        }
-      }
+      // NOTE(brentvatne): temporarily disable logging this until we can revisit the formatting
+      // and the logic for it - it's a bit too aggressive right now, and warns even if you're
+      // not using EAS Build
+      //
+      // const fingerprintsWithoutCompatibleBuilds = runtimeToCompatibleBuilds.find(
+      //   ({ runtimeVersion }) => runtimeVersion === runtime.runtimeVersion
+      // )?.fingerprintInfoGroupWithCompatibleBuilds;
+      // if (fingerprintsWithoutCompatibleBuilds) {
+      //   const missingBuilds = Object.entries(fingerprintsWithoutCompatibleBuilds).filter(
+      //     ([_platform, fingerprintInfo]) => !fingerprintInfo.build
+      //   );
+      //   if (missingBuilds.length > 0) {
+      //     const project = await AppQuery.byIdAsync(graphqlClient, projectId);
+      //     Log.warn('No compatible builds found for the following fingerprints:');
+      //     for (const [platform, fingerprintInfo] of missingBuilds) {
+      //       const fingerprintUrl = new URL(
+      //         `/accounts/${project.ownerAccount.name}/projects/${project.slug}/fingerprints/${fingerprintInfo.fingerprintHash}`,
+      //         getExpoWebsiteBaseUrl()
+      //       );
+      //       Log.warn(
+      //         formatFields(
+      //           [
+      //             {
+      //               label: `${this.prettyPlatform(platform)} fingerprint`,
+      //               value: fingerprintInfo.fingerprintHash,
+      //             },
+      //             { label: 'URL', value: fingerprintUrl.toString() },
+      //           ],
+      //           {
+      //             labelFormat: label => `    ${chalk.dim(label)}:`,
+      //           }
+      //         )
+      //       );
+      //       Log.addNewLineIfNone();
+      //     }
+      //   }
+      // }
     }
   }
 
-  private prettyPlatform(updatePlatform: string): string {
-    switch (updatePlatform) {
-      case 'android':
-        return 'Android';
-      case 'ios':
-        return 'iOS';
-      default:
-        return updatePlatform;
-    }
-  }
+  // private prettyPlatform(updatePlatform: string): string {
+  //   switch (updatePlatform) {
+  //     case 'android':
+  //       return 'Android';
+  //     case 'ios':
+  //       return 'iOS';
+  //     default:
+  //       return updatePlatform;
+  //   }
+  // }
 
   private sanitizeFlags(flags: RawUpdateFlags): UpdateFlags {
     const nonInteractive = flags['non-interactive'] ?? false;